// packages/ngx-castle-ui/src/lib/types/baseTypes.ts
export type DisplayClassType =
  | 'primary'
  | 'secondary'
  | 'accent'
  | 'success'
  | 'warn'
  | 'error'
  | 'muted'
  | 'primary-alt'
  | 'secondary-alt'
  | 'none'
  | '';

export type SizeType = 'xsmall' | 'small' | 'med' | 'medium' | 'large';
export type AlignType =
  | 'inline'
  | 'top'
  | 'top-right'
  | 'bottom'
  | 'left'
  | 'right'
  | 'center';

export type RadiusSizeType = 'small' | 'med' | 'large' | 'none';
export type UiEffectsType = 'transition' | 'rotate' | 'scale' | 'none';
export type UiEffectsAmountType = 'small' | 'med' | 'large';
export type BtnModeType = 'link' | 'click' | 'dialog' | 'modal';
export type DisplayType = 'basic' | 'flat' | 'outline' | '';
export type ModeType = 'read' | 'edit';
export type btnType = 'button' | 'submit' | 'reset';

export type StepperDetailsType = { title: string; desc?: string };

// export type fontStyleType = 'h1' | 'h2' | 'h3';